requirements.txt
# Core Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0
jinja2==3.1.6

# Data Processing & Analysis
pandas==2.2.3
numpy==2.2.1
numpy-financial==1.0.0
scipy==1.14.1
openpyxl==3.1.5

# API & Validation
pydantic==2.11.0
pydantic-settings==2.7.1
requests==2.32.3

# Database & Redshift
redshift-connector==2.1.7
python-dotenv==1.0.1
sqlalchemy==2.0.37

# Utilities
chardet==5.2.0
sse-starlette==2.2.1
pyarrow==19.0.0
tenacity==9.0.0

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
mypy==1.11.0